moved add accepted action functionality to where it is used instead of top parent
